+++ packages/block-editor/src/private-apis.js
@@ -15,7 +15,7 @@ import {
 import { PrivateListView } from './components/list-view';
 import InspectorControlsLastItem from './components/inspector-controls/last-item';
 import { useHasBlockToolbar } from './components/block-toolbar/use-has-block-toolbar';
-import { cleanEmptyObject } from './hooks/utils';
+import { cleanEmptyObject, usePrivateStyleOverride } from './hooks/utils';
 import BlockQuickNavigation from './components/block-quick-navigation';
 import { LayoutStyle } from './components/block-list/layout';
 import BlockManager from './components/block-manager';
@@ -63,6 +63,12 @@ import {
 } from './components/block-list/use-block-props/use-block-refs';
 import { LinkPicker } from './components/link-picker';
 import useRemoteUrlData from './components/link-control/use-rich-url-data';
+import { PrivateBlockContext } from './components/block-list/private-block-context';
+import useListViewPanelState from './components/use-list-view-panel-state';
+import {
+	isHashLink,
+	isRelativePath,
+} from './components/link-control/is-url-like';
 
 /**
  * Private @wordpress/block-editor APIs.
@@ -83,6 +89,7 @@ lock( privateApis, {
 	InspectorControlsLastItem,
 	useHasBlockToolbar,
 	cleanEmptyObject,
+	usePrivateStyleOverride,
 	BlockQuickNavigation,
 	LayoutStyle,
 	BlockManager,
@@ -122,4 +129,8 @@ lock( privateApis, {
 	useBlockElementRef,
 	LinkPicker,
 	useRemoteUrlData,
+	PrivateBlockContext,
+	useListViewPanelState,
+	isHashLink,
+	isRelativePath,
 } );